+++ changelog.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+-   Handles page scroll position on page load [#36](https://github.com/Pageworks/djinnjs/issues/36)
+-   Scroll to element with id if hash is in url
+
+### Fixed
+
 -   Windows filename regex issues [#28](https://github.com/Pageworks/djinnjs/issues/28)
 
 ## [0.0.14] - 2020-01-13